+++ packages/deployment-service/src/functions/pipeline/index.ts
@@ -11,4 +11,4 @@ export * from './pipeline-runner-paginate'
 export * from './codebuild-pipeline-updater'
 export * from './codebuild-deploy'
 export * from './pipeline-setup'
-export * from './pipeline-trear-down'
+export * from './pipeline-tear-down'
